pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Artur Araujo <arturavs1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.115.5"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.17"
httpcore = "^1.0.7"
pyjwt = "^2.10.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
taskipy = "^1.14.0"

[tool.pytest.ini_options]
pythonpath = "."
addpts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix &&  ruff format .'
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

[tool.ruff]
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']
ignore = ["E501"]

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
